pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ko2cube"
version = "0.1.0"
description = "ko2cube project"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.3",
    "fastapi>=0.115.0",
    "uvicorn>=0.24.0",
    "kubernetes>=29.0.0",
    "PyYAML>=6.0.1",
    "pytest>=8.0.0",
]

[tool.setuptools.packages.find]
where = [".", "server"]

[project.scripts]
server = "server.app:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: marks tests as integration tests requiring a live kwok cluster (deselect with '-m not integration')",
]
